+++ QCamera2/stack/mm-camera-interface/src/mm_camera_thread.c
@@ -44,7 +44,7 @@
 typedef enum {
     /* poll entries updated */
     MM_CAMERA_PIPE_CMD_POLL_ENTRIES_UPDATED,
-    /* poll entries updated */
+    /* poll entries updated asynchronous */
     MM_CAMERA_PIPE_CMD_POLL_ENTRIES_UPDATED_ASYNC,
     /* commit updates */
     MM_CAMERA_PIPE_CMD_COMMIT,
@@ -92,16 +92,18 @@ static int32_t mm_camera_poll_sig_async(mm_camera_poll_thread_t *poll_cb,
     pthread_mutex_lock(&poll_cb->mutex);
     /* reset the statue to false */
     poll_cb->status = FALSE;
-    /* send cmd to worker */
 
+    /* send cmd to worker */
     ssize_t len = write(poll_cb->pfds[1], &cmd_evt, sizeof(cmd_evt));
-    if(len < 1) {
-        CDBG_ERROR("%s: len = %zd, errno = %d", __func__, len, errno);
+    if (len < 1) {
+        CDBG_ERROR("%s: len = %lld, errno = %d", __func__,
+                (long long int)len, errno);
         /* Avoid waiting for the signal */
         pthread_mutex_unlock(&poll_cb->mutex);
         return 0;
     }
-    CDBG("%s: begin IN mutex write done, len = %zd", __func__, len);
+    CDBG("%s: begin IN mutex write done, len = %lld", __func__,
+            (long long int)len);
     pthread_mutex_unlock(&poll_cb->mutex);
     CDBG("%s: X", __func__);
     return 0;
@@ -140,12 +142,14 @@ static int32_t mm_camera_poll_sig(mm_camera_poll_thread_t *poll_cb,
 
     ssize_t len = write(poll_cb->pfds[1], &cmd_evt, sizeof(cmd_evt));
     if(len < 1) {
-        CDBG_ERROR("%s: len = %zd, errno = %d", __func__, len, errno);
+        CDBG_ERROR("%s: len = %lld, errno = %d", __func__,
+                (long long int)len, errno);
         /* Avoid waiting for the signal */
         pthread_mutex_unlock(&poll_cb->mutex);
         return 0;
     }
-    CDBG("%s: begin IN mutex write done, len = %zd", __func__, len);
+    CDBG("%s: begin IN mutex write done, len = %lld", __func__,
+            (long long int)len);
     /* wait till worker task gives positive signal */
     if (FALSE == poll_cb->status) {
         CDBG("%s: wait", __func__);
